Продолжаю исправлять микрофронтэнд cardboard
